Disallow more than 1 public method named with regex
Right now it is possible to define a class that has two public methods:
`exampleOne` and `exampleTwo`. If you do not give this class extra
methods and configure the ShouldHaveOnlyOnePublicMethodNamed rule to
look for `^example.+$` then it will report success, despite two methods
existing with this naming scheme.

This rule will only start failing after introducing a method which does
*not* follow the naming scheme, e.g. `foo`. This PR aims to fix this by
also failing in the `exampleOne` and `exampleTwo` scenario.

Note that we interpreted this rule as "can have several public methods,
but only one named ...", instead of "should have only one public method
*and* it is named ...", so this might not follow the original intention
of the rule.